Se crean las rutas para cada pagina
