+++ package.json
@@ -10,6 +10,7 @@
     "bootstrap-icons": "^1.8.2",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
