+++ src/Pages/Prueba.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const Prueba = () => {
+  return <div>Prueba</div>;
+};
